「最近のスコア」機能を小数点スコアに対応

+++ db/migrate/20241231073509_create_recent_rank_items.rb
@@ -2,7 +2,7 @@ class CreateRecentRankItems < ActiveRecord::Migration[7.1]
   def change
     create_table :recent_rank_items do |t|
       t.references :ranking_board, null: false, foreign_key: true
-      t.integer :score, null: false, index: true, comment: "最新の記録"
+      t.float :score, null: false, index: true, comment: "最新の記録"
       t.string :uid, null: false
       t.boolean :new_record_score, null: false, default: false, comment: "新記録かどうか"
